✅ TARGET 3 COMPLETE: Async/Sync Duplication Elimination
🎯 MASSIVE ACHIEVEMENT - Week 1 Technical Debt Elimination Complete!

ELIMINATED:
• ModularPDFExtractor.swift (671 lines) → Kept AsyncModularPDFExtractor.swift
• All DispatchSemaphore usage (0 remaining in codebase)
• Sync PDF processing services → Async-first architecture achieved
• Mock services from main app target → Moved to test target

TECHNICAL IMPROVEMENTS:
• AsyncModularPDFExtractor now conforms to PDFExtractorProtocol
• DI containers updated to use async services only
• Error type mappings fixed after ModularExtractionError deletion
• Clean async/await patterns throughout codebase

BUILD STATUS: ✅ SUCCESS
Lines Eliminated This Target: 671+
Total Week 1 Elimination: ~4,500+ lines

WEEK 1 TARGETS COMPLETE:
✅ Target 1: Text Extraction Massacre (2000+ lines)
✅ Target 2: Test Infrastructure Purge (2000+ lines, 26+ files)
✅ Target 3: Async/Sync Duplication Elimination (671+ lines)

Ready for Week 2: Complexity Reduction Phase 🚀

+++ PayslipMax/Services/Extraction/ExtractionStrategySelector.swift
@@ -165,8 +165,8 @@ class ExtractionStrategySelector: ExtractionStrategySelectorProtocol {
         let memoryPressure = Double(estimatedMemoryNeed) / Double(availableMemory)
         
         return ResourceAssessment(
-            availableMemoryMB: availableMemory / (1024 * 1024),
-            estimatedMemoryNeedMB: estimatedMemoryNeed / (1024 * 1024),
+            availableMemoryMB: UInt64(availableMemory) / (1024 * 1024),
+            estimatedMemoryNeedMB: UInt64(estimatedMemoryNeed) / (1024 * 1024),
             memoryPressureRatio: memoryPressure,
             processorCoreCount: ProcessInfo.processInfo.processorCount
         )
